Using ethers to print the transaction hash
Attempting to mock out the whole webautn flow was far too complicated,
so introducing some offline steps that require copy+paste with the devs
helps keep the code close to the gateway without actually using it.

Attempting to use the exact viem code in the gateway was harder than
finding an example ethers use of the smart contract helpers. Building
the transaction in ethers still requires lots of boilerplate, but at
least the transaction type is better supported!