Update spend limit module to be more user-friendly
Everything upstream of this is now broken for init, and actual
validation and timestamp checking is still disconnected, but the data is
setup in a nicer way.

This also removes the spend-limit per passkey in exchange for a global
spend limit, which is easier for now, but we'll want the per-key limits
back later.

Links to #45