fix: use ecrecover within spendlimit module
Update the init arguments to take an array, as now we have 2 possible
validators. It's not clear what the other module type does now, so would
be good to consolidate and use the module type system for a single array
that gets split by type.

Haven't yet been able to test this signing, but it might work!