Session Validation Module Interface Update (#208)
* chore: clean up unused code

R1 validation and keys were unreachable after changing the init
interface,
K1 validation was unused since it was validated directly without a
module (simple signature instead of modular signature)
Fallback modules were entirely unused.
Modules and exec modules could have been used, but were not and weren't
supported by the factory to be installed at init.

* fix: interface size

dropping unused arg

* fix: single module interface

Moving to a single module to elimiate duplicate code

* fix: replace modules with hooks

This is much more direct path to installing instead of using the install
to re-enter the account

* fix: run all tests

removing only locally tests pass

* fix: revert this to older implemention

This broke locally, but now fails in CI, maybe the converse is also true

* chore: cleanup test code

This wasn't necessary to pass in ci, but now fails locally :(

* fix: await on test transaction failure

* fix: have to await the reverted within expect

This is just a test setup issue combined with a era-test-node change

* fix: no validation hook

Remove clave interface for validation, only need one interface

* feat: move to two interfaces for module validation

Would love to have a single interface, but 1271 makes the extra
transaction argument difficult, so an easy workaround is just having two

Having them named the same is also slightly confusing when one is only
called from isValidSignature

* feat: remove old validator interface

So now there are only validation modules, k1 owners, and execution
hooks. And only validation modules and k1 owners can be installed on
creation.

* feat: re-add execution hook

also rename validation functions.

Still trying to understand how the data is different without hook
interface.

* fix: gas estimation during session validation

Couldn't find a better way to do this and I don't know how this wasn't a
problem before?

* fix: more merge conflicts

* chore: cleanup accidental changes

To reduce noise in diff

* feat: re-allow validation type hooks

These aren't being used, but are added for some unique flavor in case
someone can find a use case.
Removing the hook data concept as it's difficult to manage if the signer
isn't aware of all of the hooks on the account.

* fix: validation hook missing lines

* fix: removed extra hash for gas estimates

This turned out to just cost more gas with no benefit to estimation
accuracy.

* fix: cleanup

* chore: format

* fix: remove invalid test

* fix: more fixes

* fix: gas estimation

* fix: add session during init

---------

Co-authored-by: Lyova Potyomkin <lyova.potyomkin@gmail.com>