Passkey tests (#231)
* feat: create an independant implemention for signing

Instead of having the browser in the loop, we can just sign this
ourselves.
Just need to figure out the formats ;)

* feat: off-chain verification test

Something must be getting mixed up, because these recorded signatures
already passed the simplewebauth verification checks so they should
verify with the web crpyto libs.

Will need to dig through that lib again to see what's going on compared
to how I recorded these.

* chore: refactor to compare as directly as possible

So there was issue where the signature was too long and it was hard to
see, but now all the data formats are using raw bytes so should be
directly compable.

No clue as to what's incompatible between the two ES256 implementions

* feat: add more generated signing tests

The problem before was double hash,
the problem now is making bad s values during key generation

* feat: refactor test coverage

JSON parsing fails on duplicate keys, so it's impossible to get coverage
on the duplicate key checks.
Also explicitly not testing cross-origin, as we don't care anymore about
it's value (unless we want to save that with the key, because it should
otherwise be managed by the signer with the rp hosts)

* feat: validate client json key order

Don't allow any invalid keys instead of just checking the most recent
key in case there are duplicate keys in the JSON

* feat: add r max check

Rename the variables and have the same range check as low S

* chore: use only unnamed returns

* Update test/PasskeyModule.ts

Co-authored-by: Nicolas Villanueva <1890113+MexicanAce@users.noreply.github.com>

* fix: use viem instead of ethers for key encoding

Helps with consistency between the SDK and contract tests, also helps
clean up the test code with a quick refactor

---------

Co-authored-by: Nicolas Villanueva <1890113+MexicanAce@users.noreply.github.com>
Co-authored-by: Lyova Potyomkin <lyova.potyomkin@gmail.com>