JSON Parser Test Suite (#232)
* feat: port tests from jsmnsol lib

The truffle tests were pretty minimal, but good to keep in sync since
this isn't a published package we're importing.
https://github.com/chrisdotn/jsmnSol/tree/master/test

* feat: add jsc JSONTestSuite

Since the existing test suite didn't catch any of the reported problems
with the JSON parsing, adding another larger suite that's used to
compare popular JSON parsers should!

Obviously because this is so large it can be trimmed, but the hope is to
find overlap between the feedback and the test-cases here.

* fix: update to latest era-test-node

This action is broken again, hopefully the latest works!

* fix: provide more debug for json tests

Tests connect to the files

* fix: parse an emoji instead of backslash

Skipping all the json format tests for now,
These are helpful if we want a fully compliant parser, but aren't
helpful in parsing the client data JSON.

* feat: add duplicate key parsing test

It doesn't return an error, because it doesn't actually record key
values

* fix: undo era-test-node version

it looks like this is fixed now

* Update test/JsmnSolLib.ts

Co-authored-by: Nicolas Villanueva <1890113+MexicanAce@users.noreply.github.com>

* Update test/JsmnSolLib.ts

Co-authored-by: Nicolas Villanueva <1890113+MexicanAce@users.noreply.github.com>

* Update test/JsmnSolLib.ts

Co-authored-by: Nicolas Villanueva <1890113+MexicanAce@users.noreply.github.com>

* feat: add json string tests

Covers existing functionality

* fix: re-add removed return result

---------

Co-authored-by: Nicolas Villanueva <1890113+MexicanAce@users.noreply.github.com>
Co-authored-by: Lyova Potyomkin <lyova.potyomkin@gmail.com>