Update module install and remove interfaces (#238)
* feat: remove disable and init from module

Modules no longer have any permission to act on the account, so any
disabling will need to come from the account, removing the last use case
for disable.

Init was functionality duplicated with the add key functionality, so it
was removed in favor of the existing interface.

* feat: add event to adding a new passkey

Publishes the domain since that's useful vs the raw bytes which is hard

* fix: update tests with new interface

The revert if exists functionality moves to the account
Also the disable functions are removed again

* feat: use 7579 module interface for add/remove

Adds a new entry point so we can remove modules

* fix: add IModule to interface support

A little scary that the webauthn module didn't throw errors, so shows
the need for some more tests for that module install

* feat: add tests to cover passkey init

This covers the initial flow that got broken when the module didn't
support the expected interface

* fix: use IModule for addHook interface

This no longer needs to be custom because there's no hooks yet

* chore: remove unused lib

Have different options to remove trusted vs untrusted modules

* fix: remove vs unlink

* fix: typo

---------

Co-authored-by: Nicolas Villanueva <1890113+MexicanAce@users.noreply.github.com>
Co-authored-by: Lyova Potyomkin <lyova.potyomkin@gmail.com>