Store Credential id with passkey (#270)
* feat: store and parse credential id from passkey

The odds that the parsing are correct is basically zero,
but right now I can't even build this locally so hopefully CI has better luck?

* fix: compile error

* fix: provide credential id externally

Parsing it from the auth data would only work on creation, not on get

* fix: don't update hardhat

can't figure out how to update pnpm lockfile

* feat: dedicated revoke passkey

Also allows checking if the passkey exists so you can avoid
adding the same passkey to multiple accounts on the same domain

* chore: revert hardhat upgrade

Annoying to do between two repos with the lockfile

* feat: create lookup to prevent accidential duplicate keys

For cross-origin support it would be possible to try the same
key on another domain.
You also might want to check if you have an account already without
knowing your address and just your passkey

* chore: upgrade zksolc

* feat: deploy script saves addresses to file

This should help sync the contracts to the SSO repo when interating
on contracts

* feat: add checks to prevent overwriting

Without the extra layer it would have been possible to hide
someone's key in the map or to make it look like the map did exist
when it shouldn't have.

* fix: remove .only from passkey tests

leftover from debugging

* feat: simplify key storage

Use a single mapping instead of multiple
For whatever reason, the auto-generated view function includes an extra argument
So there's a dedicated view function that avoids it

* docs: update comment to explain new mapping

This is entirely related to the factory unique id and how SSO populates it by default with the
credential id, so it's not clear from only the contracts repo how this is useful. The come
k# Please enter the commit message for your changes. Lines starting

* feat: merge mappings into single struct

Now that AA validation allows for more than single value mapping access
We can simplify our storage layout so we store the key and address together

The types of the autogenerated view function don't work correctly,
so adding a dedicated view function fixes this

* feat: add script to publish abi

This should avoid the ABI falling out of sync from having
to manually re-create it.
The formatting is a little rough, but should work?

* fix: back to two mappings

The validation logic isn't quite smart enough to figure out if the
first part of the struct is the account address, so going back to
two mappings

* fix: add as const to publish export

This is the trick to getting viem to read the type correctly

* fix: merge conflict

* fix: code review comments

Update names, change remove to memory and revert on failure

---------

Co-authored-by: Lyova Potyomkin <lyova.potyomkin@gmail.com>