WebAuthN Audit Changes (#333)
* chore: set webauthn public keys to private

Add some comments to explain what's going on within the code,
as the past was only listed in the commit messages and PRs.

Also update the contract layout, and add error messages for use later

* chore: update key add to custom errors

Instead of returning false (which didn't really work anyways),
this function now reverts with custom errors that we can individually
validate.

Also removes some abstractions that we don't need for testing
(VerfifierCaller was used pre-precompile for testing)

* chore: pre-hash string comparison

Slight decrease in gas usage! (before: 243108, after: 242376)
N-19

* chore: add key length validation for passkey metadata

L-06 to check that the data is valid

* chore: avoid extra storage access during key removal

From N-03, a small improvement

* chore: add doc to explain what module means

There's no dependancy management in 7579

* chore: move mapping into state var section

What a waste